Added custom exceptions
