+++ server/BLL/Exceptions/Auth/FailedSignupException.cs
@@ -1,0 +1,8 @@
+﻿namespace BLL.Exceptions.Auth;
+
+public class FailedSignupException: Exception
+{
+    public FailedSignupException(string message): base($"Failed to sign you up: {message}")
+    {
+    }
+}
