+++ server/BLL/Exceptions/Auth/InvalidLoginException.cs
@@ -1,0 +1,8 @@
+﻿namespace BLL.Exceptions.Auth;
+
+public class InvalidLoginException: Exception
+{
+    public InvalidLoginException(): base("Invalid email or password")
+    {
+    }
+}
