+++ server/BLL/MappingProfiles/UserProfile.cs
@@ -1,0 +1,6 @@
+﻿namespace BLL.MappingProfiles;
+
+public class UserProfile
+{
+    
+}
